investment_assistant/requirements.txt
requests
beautifulsoup4
yfinance
python-telegram-bot
google-generativeai
python-dotenv
